pyproject.toml
[project]
name = "frc-regwatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitdb>=4.0.12",
    "gitpython>=3.1.45",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
